+++ sdk/core/azure-core/src/context.cpp
@@ -11,15 +11,17 @@ Context& Azure::Core::Context::GetApplicationContext()
   return context;
 }
 
-Azure::DateTime Azure::Core::Context::GetExpiration() const
+Azure::DateTime Azure::Core::Context::GetDeadline() const
 {
+  // Contexts form a tree. Here, we walk from a node all the way back to the root in order to find
+  // the earliest deadline value.
   auto result = DateTime::max();
   for (auto ptr = m_contextSharedState; ptr; ptr = ptr->Parent)
   {
-    auto expiration = ContextSharedState::FromDateTimeRepresentation(ptr->Expiration);
-    if (result > expiration)
+    auto deadline = ContextSharedState::FromDateTimeRepresentation(ptr->Deadline);
+    if (result > deadline)
     {
-      result = expiration;
+      result = deadline;
     }
   }