+++ sdk/core/azure-core/src/http/retry_policy.cpp
@@ -161,7 +161,7 @@ Context inline CreateRetryContext(Context const& parent)
   {
     retryCount = parent.GetValue<int>(RetryKey) + 1;
   }
-  return parent.CreateWithValue(RetryKey, retryCount);
+  return parent.CreateChildContext(RetryKey, retryCount);
 }
 } // namespace