+++ sdk/storage/azure-storage-common/inc/azure/storage/common/storage_switch_to_secondary_policy.hpp
@@ -15,7 +15,7 @@ namespace Azure { namespace Storage { namespace _internal {
 
   inline Azure::Core::Context WithReplicaStatus(const Azure::Core::Context& context)
   {
-    return context.CreateWithValue(SecondaryHostReplicaStatusKey, std::make_shared<bool>(true));
+    return context.CreateChildContext(SecondaryHostReplicaStatusKey, std::make_shared<bool>(true));
   }
 
   class StorageSwitchToSecondaryPolicy : public Azure::Core::Http::Policies::HttpPolicy {