+++ sdk/core/azure-core/CHANGELOG.md
@@ -16,8 +16,8 @@
   - `GetMinorVersion`.
 - Changes to `Azure::Core::Context`:
   - Renamed `Get()` to `GetValue()`.
-  - Renamed `WithValue()` to `CreateWithValue()`.
-  - Renamed `WithDeadline()` to `CreateWithExpiration()`, changed input parameter type to `Azure::DateTime`.
+  - Renamed `WithValue()` to `CreateChildContext()`.
+  - Renamed `WithDeadline()` to `CreateChildContext()`, and changed input parameter type to `Azure::DateTime`.
 - Removed from `Azure::Core::Http::Policies` namespace: `HttpPolicyOrder`, `TransportPolicy`, `RetryPolicy`, `RequestIdPolicy`, `TelemetryPolicy`, `BearerTokenAuthenticationPolicy`, `LogPolicy`.
 - Renamed `Azure::Core::Http::RawResponse::GetBodyStream()` to `ExtractBodyStream()`.
 
@@ -29,7 +29,7 @@
 - Added `Azure::Core::Operation<T>::GetRawResponse()`.
 - Added `Azure::Core::PackageVersion`.
 - Added support for logging to console when `AZURE_LOG_LEVEL` environment variable is set.
-- Added `Azure::Core::Context::GetExpiration()`.
+- Added `Azure::Core::Context::GetDeadline()` and `Azure::Core::Context::HasDeadline()`.
 
 ### Breaking Changes